python-bindings: Add wrapper for download-error
Since download-error emits a GError, which pygobject doesn't wrap,
the dowload-error signal was useless. The new method on_error connects
to the signal for the python object, and converts the result when it
is triggered.

Needs libpyglib, which does not have a pkg-config file, and is
therefore linked to in a very hackish way.