download: Add download state info, pause ability
Replace s_d_get/set_completed with s_d_get_state. State is read-only,
and controlled by signals. Add download-done signal to be able to set
the state to done.

Add code to make it possible to pause downloads. Libsoup does not yet
support pause/resume, so it is only implemented for torrents.